+++ apps/demo/src/app/form/form-data.pipe.ts
@@ -1,0 +1,17 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({ name: 'appFormData' })
+export class FormDataPipe implements PipeTransform {
+  transform(formData: FormData): { key: string; name: string }[] {
+    if (!formData) {
+      return undefined;
+    }
+
+    const result = [];
+    formData.forEach((value, key) => {
+      const { name } = value as File;
+      result.push({ key, name });
+    });
+    return result;
+  }
+}