+++ libs/core/src/lib/dynamic-form-validation/dynamic-form-validation-error.ts
@@ -1,3 +1,4 @@
 export interface DynamicFormValidationError {
   message?: string;
+  [key: string]: any;
 }